feat(native): More concise module structure
